+++ interfaces/src/main/java/ar/edu/itba/paw/service/ImageService.java
@@ -3,11 +3,11 @@ package ar.edu.itba.paw.service;
 import java.util.Optional;
 
 public interface ImageService {
-    long create(byte[] bytes);
+    int create(byte[] bytes);
 
-    Optional<byte[]> getById(long imageId);
+    Optional<byte[]> getById(int imageId);
 
-    boolean update(long imageId, byte[] bytes);
+    boolean update(int imageId, byte[] bytes);
 
-    boolean delete(long imageId);
+    boolean delete(int imageId);
 }
