+++ models/src/main/java/ar/edu/itba/paw/model/User.java
@@ -2,13 +2,13 @@ package ar.edu.itba.paw.model;
 
 public class User {
 
-    private final long userId;
+    private final int userId;
     private final String username;
     private final String password;
     private final String name;
     private final String email;
 
-    public User(long userId, String username, String password, String name, String email) {
+    public User(int userId, String username, String password, String name, String email) {
         this.userId = userId;
         this.username = username;
         this.password = password;
@@ -32,7 +32,7 @@ public class User {
         return email;
     }
 
-    public long getUserId() {
+    public int getUserId() {
         return userId;
     }
 }
